Adding getopts and option to not remove images

+++ extended-rancher-2-cleanup/extended-cleanup-rancher2.sh
@@ -2,9 +2,33 @@
 # Backup your data
 # Use at your own risk
 # Usage ./extended-cleanup-rancher2.sh
-# Include clearing all iptables: ./extended-cleanup-rancher2.sh flush
+
+FlushIptables="false"
+RemoveImages="false"
+while getopts ":f:i:" opt; do
+  case $opt in
+    f)
+      FlushIptables="true"
+      ;;
+    i)
+      RemoveImages="true"
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      exit 1
+      ;;
+    :)
+      echo "Option -$OPTARG requires an argument." >&2
+      exit 1
+      ;;
+  esac
+done
+
 docker rm -f $(docker ps -qa)
-docker rmi -f $(docker images -q)
+if [[ "$RemoveImages" == "true" ]]
+then
+  docker rmi -f $(docker images -q)
+fi
 docker volume rm $(docker volume ls -q)
 for mount in $(mount | grep tmpfs | grep '/var/lib/kubelet' | awk '{ print $3 }') /var/lib/kubelet /var/lib/rancher; do umount $mount; done
 cleanupdirs="/etc/ceph /etc/cni /etc/kubernetes /opt/cni /opt/rke /run/secrets/kubernetes.io /run/calico /run/flannel /var/lib/calico /var/lib/weave /var/lib/etcd /var/lib/cni /var/lib/kubelet /var/lib/rancher/rke/log /var/log/containers /var/log/pods /var/run/calico"
@@ -17,7 +41,7 @@ for interface in $cleanupinterfaces; do
   echo "Deleting $interface"
   ip link delete $interface
 done
-if [ "$1" = "flush" ]; then
+if [ "$FlushIptables" = "true" ]; then
   echo "Parameter flush found, flushing all iptables"
   iptables -F -t nat
   iptables -X -t nat
